V1 GestionReservable

+++ TP4/Models/Reservation.cs
@@ -6,7 +6,7 @@
         public DateTime DateDebut { get; set; }
         public DateTime DateFin { get; set; }
         public IReservable ObjetDeLaReservation { get; set; }
-        public double Prix { get; set; }
+        public int Prix { get; set; }
         public int PrixJournalier { get; set; }
 
         public Reservation(int id, DateTime dateDebut, DateTime dateFin, IReservable objetDeLaReservation)
